refactor(#97): vehicles/stores — bulletproof-react segmentation
git mv vehicleFormState{,.test}.ts → stores/.
External call site updated: VehicleCreatePage (canSubmit).
