+++ src/pages/VehicleCreatePage.tsx
@@ -10,7 +10,7 @@ import DiscardDialog from '../components/dialogs/DiscardDialog';
 import { BLANK_FORM } from '../data/vehicles/utils/vehicleFormDefaults';
 import { useVehiclePairSave } from '../data/vehicles/xml/useVehiclePairSave';
 import { useDirtyFormBlock } from '../data/vehicles/useDirtyFormBlock';
-import { canSubmit } from '../data/vehicles/vehicleFormState';
+import { canSubmit } from '../data/vehicles/stores/vehicleFormState';
 import { waitForVehicleInList } from '../data/vehicles/api/waitForVehicleInList';
 import { vehicleSelectedHref } from '../data/vehicles/utils/vehicleUrlParams';
 import { useAuth } from '../auth/authUtils';
